Added fir filter example

+++ buffer.c
@@ -4,7 +4,7 @@
 
 
 void initBuffer(buffer* buffer){
-	buffer->buff = (int *) malloc(BUFF_SIZE * sizeof(int));
+	buffer->buff = (short *) malloc(BUFF_SIZE * sizeof(short));
 	int i;
 	buffer->index = 0;
 	for(i=0;i<BUFF_SIZE;i++){
@@ -12,14 +12,14 @@ void initBuffer(buffer* buffer){
 	}
 }
 
-void push(buffer* buffer,float value){
+void push(buffer* buffer,short value){
 	buffer->buff[(buffer->index++) & BUFF_SIZE_MASK] = value;
 }
 
-float pop(buffer* buffer){
+short pop(buffer* buffer){
 	return buffer->buff[(buffer->index + (~BUFF_SIZE_MASK)) & BUFF_SIZE_MASK];
 }
 
-float readn(buffer* buffer, int Xn){
+short readn(buffer* buffer, int Xn){
 	return buffer->buff[(buffer->index + (~Xn)) & BUFF_SIZE_MASK];
 }
